python-scripts/invoice-system/requirements.txt
anthropic
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
reportlab
python-dotenv
